+++ demo/cypress/e2e/navigation.cy.js
@@ -1,55 +1,147 @@
 describe('Navigation', () => {
-  beforeEach(() => {
-    cy.visitComponentUrl('header/with_navigation');
-    cy.viewport('iphone-6');
+  context('when on a large screen', () => {
+    beforeEach(() => {
+      loadComponentExample();
+      cy.viewport(1440, 860);
+    });
+
+    it('renders navigation with no greedy navigation', () => {
+      cy.findByRole('button', { name: /More/i }).should('not.exist');
+      cy.get('.cads-greedy-nav').should('have.attr', 'aria-haspopup', 'false');
+    });
   });
 
-  it('navigation links can be viewed by opening/closing the dropdown', () => {
-    openNavigation();
-    assertNavigationOpen();
-    closeNavigation();
-    assertNavigationClosed();
+  context('when greedy navigation is first triggered', () => {
+    beforeEach(() => {
+      loadComponentExample();
+      cy.viewport(800, 600);
+    });
+
+    it('has expected aria attributes', () => {
+      cy.get('.cads-greedy-nav').should('have.attr', 'aria-haspopup', 'true');
+      cy.findByRole('button', { name: /More/i }).should(
+        'have.attr',
+        'aria-controls',
+        'cads-greedy-nav-dropdown',
+      );
+    });
+
+    it('moves navigation items into the greedy navigation', () => {
+      cy.findByRole('button', { name: /More/i }).click();
+      assertItemsInMainNavigation([
+        'Benefits',
+        'Work',
+        'Debt and money',
+        'Consumer',
+        'Housing',
+        'Family',
+        'Law and courts',
+      ]);
+      assertItemsInGreedyNavigation(['Immigration', 'Health', 'More from us']);
+    });
   });
 
-  it('tabbing into the dropdown toggle automatically opens the dropdown menu', () => {
-    tabIntoNavigation();
-    assertNavigationOpen();
+  context('when on a small screen', () => {
+    beforeEach(() => {
+      loadComponentExample();
+      cy.viewport(375, 667);
+    });
+
+    it('moves navigation items into the greedy navigation including header links', () => {
+      cy.findByRole('button', { name: /More/i }).click();
+      assertItemsInMainNavigation(['Benefits', 'Work']);
+
+      assertItemsInGreedyNavigation([
+        'Debt and money',
+        'Consumer',
+        'Housing',
+        'Family',
+        'Law and courts',
+        'Immigration',
+        'Health',
+        'More from us',
+        'AdviserNet',
+        'Cymraeg',
+        'Sign in',
+      ]);
+    });
   });
 
-  // NP-1755
-  it.skip('tabbing out of the dropdown menu automatically closes it');
+  context('when interacting with a mouse', () => {
+    beforeEach(() => {
+      loadComponentExample();
+      cy.viewport(375, 667);
+    });
 
-  it('can close the dropdown menu using your keyboard', () => {
-    openNavigation().type('{esc}');
-    assertNavigationClosed();
+    it('allows opening and closing the greedy navigation', () => {
+      cy.findByRole('button', { name: /More/i }).click();
+      assertNavigationOpen();
+      cy.findByRole('button', { name: /Close/i }).click();
+      assertNavigationClosed();
+    });
+
+    it('closes the greedy navigation when clicking outside it', () => {
+      cy.findByRole('button', { name: /More/i }).click();
+      assertNavigationOpen();
+      cy.get('body').click();
+      assertNavigationClosed();
+    });
   });
 
-  it('can close the dropdown menu by clicking outside it', () => {
-    openNavigation();
-    assertNavigationOpen();
-    cy.get('body').click();
-    assertNavigationClosed();
+  // Note: Tab interaction is a bit fraught to test in Cypress
+  // We use https://github.com/dmtrKovalenko/cypress-real-events
+  // to approximate this but full keyboard testing should still be
+  // done manually when changing this behaviour.
+  context('when interacting with a keyboard', () => {
+    beforeEach(() => {
+      loadComponentExample();
+      cy.viewport(375, 667);
+    });
+
+    it('closes the greedy navigation when tabbing out', () => {
+      cy.findByRole('button', { name: /More/i }).click();
+      cy.findByTestId('cads-greedy-nav-dropdown').within(() => {
+        cy.findByText('Sign in').focus().realPress('Tab');
+        assertNavigationClosed();
+      });
+    });
+
+    it('closes the greedy navigation when pressing the escape key', () => {
+      cy.findByRole('button', { name: /More/i }).click().type('{esc}');
+      assertNavigationClosed();
+    });
   });
 
-  function openNavigation() {
-    return cy
-      .findByText('More')
-      .should('be.visible')
-      .click()
-      .should('have.attr', 'aria-expanded', 'true')
-      .should('have.attr', 'aria-controls', 'greedy-nav-dropdown');
-  }
+  context('when in english', () => {
+    beforeEach(() => {
+      loadComponentExample('en');
+      cy.viewport(375, 667);
+    });
 
-  function closeNavigation() {
-    return cy
-      .findByText('Close')
-      .should('be.visible')
-      .click()
-      .should('have.attr', 'aria-expanded', 'false');
-  }
+    it('has translated labels when viewing in english locale', () => {
+      cy.findByRole('button', { name: /More/i })
+        .should('have.attr', 'aria-label', 'More navigation items')
+        .click();
+      cy.findByRole('button', { name: /Close/i }).click();
+    });
+  });
 
-  function tabIntoNavigation() {
-    cy.findByText('More').should('be.visible').focus().tab().tab();
+  context('when in welsh', () => {
+    beforeEach(() => {
+      loadComponentExample('cy');
+      cy.viewport(375, 667);
+    });
+
+    it('has translated labels when viewing in welsh locale', () => {
+      cy.findByRole('button', { name: /Mwy/i }).click();
+      assertNavigationOpen();
+      cy.findByRole('button', { name: /Cau/i }).click();
+      assertNavigationClosed();
+    });
+  });
+
+  function loadComponentExample(locale = 'en') {
+    cy.visitComponentUrl('header/with_navigation', locale);
   }
 
   function assertNavigationOpen() {
@@ -58,5 +150,21 @@ describe('Navigation', () => {
 
   function assertNavigationClosed() {
     cy.findByText('More from us').should('not.be.visible');
+  }
+
+  function assertItemsInMainNavigation(expected) {
+    const items = [];
+    cy.get('.cads-navigation > ul a:visible').each(($el) =>
+      items.push($el.text()),
+    );
+    cy.wrap(items).should('deep.equal', expected);
+  }
+
+  function assertItemsInGreedyNavigation(expected) {
+    const items = [];
+    cy.get('.cads-greedy-nav > ul a:visible').each(($el) =>
+      items.push($el.text()),
+    );
+    cy.wrap(items).should('deep.equal', expected);
   }
 });